ref: move gitlab testutils to fixtures
