+++ tests/sentry/integrations/gitlab/test_client.py
@@ -4,12 +4,11 @@ from unittest import mock
 import pytest
 import responses
 
+from fixtures.gitlab import GitLabTestCase
 from sentry.auth.exceptions import IdentityNotValid
 from sentry.models import Identity
 from sentry.shared_integrations.exceptions import ApiError
 from sentry.utils import json
-
-from .testutils import GitLabTestCase
 
 GITLAB_CODEOWNERS = {
     "filepath": "CODEOWNERS",
