+++ tests/sentry/integrations/gitlab/test_search.py
@@ -3,9 +3,8 @@ from urllib.parse import parse_qs
 import responses
 from django.urls import reverse
 
+from fixtures.gitlab import GitLabTestCase
 from sentry.utils import json
-
-from .testutils import GitLabTestCase
 
 
 class GitlabSearchTest(GitLabTestCase):
